Add dedicated file for DocSelection
