+++ include/vrv/toolkit.h
@@ -13,6 +13,7 @@
 //----------------------------------------------------------------------------
 
 #include "doc.h"
+#include "docselection.h"
 #include "view.h"
 
 //----------------------------------------------------------------------------
